// src/app/store/reducers/types.ts
export interface OwnerData {
    avatar_url: string
    events_url: string
    followers_url: string
    following_url: string
    gists_url: string
    gravatar_id: string
    html_url: string
    id: number
    login: string
    node_id: string
    organizations_url: string
    received_events_url: string
    repos_url: string
    site_admin: boolean
    starred_url: string
    subscriptions_url: string
    type: string
    url: string
}

export interface Item {
    id: number
    name: string
    allow_forking: true
    archive_url: string
    archived: false
    assignees_url: string
    blobs_url:string
    branches_url: string
    clone_url: string
    collaborators_url: string
    comments_url: string
    commits_url: string
    compare_url: string
    contents_url: string
    contributors_url: string
    created_at: string
    default_branch: string
    deployments_url: string
    description: string
    disabled: boolean
    downloads_url: string
    events_url: string
    fork: boolean
    forks: number
    forks_count: number
    forks_url: string
    full_name: string
    git_commits_url: string
    git_refs_url: string
    git_tags_url: string
    git_url: string
    has_discussions: false
    has_downloads: true
    has_issues: true
    has_pages: false
    has_projects: boolean
    has_wiki: boolean
    homepage: any
    hooks_url: string
    html_url: string
    is_template: boolean
    issue_comment_url: string
    issue_events_url: string
    issues_url: string
    keys_url: string
    labels_url: string
    language: string
    languages_url: string
    license: any
    merges_url: string
    milestones_url: string
    mirror_url: any
    node_id: string
    notifications_url: string
    open_issues: number
    open_issues_count: number
    owner: OwnerData
    private: false
    pulls_url: string
    pushed_at: string
    releases_url: string
    score: number
    size: number
    ssh_url: string
    stargazers_count: 767
    stargazers_url: string
    statuses_url: string
    subscribers_url: string
    subscription_url: string
    svn_url: string
    tags_url: string
    teams_url: string
    topics: string[]
    trees_url: string
    updated_at: string
    url: string
    visibility: string
    watchers: number
    watchers_count: number
    web_commit_signoff_required: boolean
}

export interface FetchItemsResponse {
    incomplete_results: boolean
    items: Item[]
    total_count: number
}